% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamify.R
\name{calc_loss_and_grads}
\alias{calc_loss_and_grads}
\title{Calculate Loss & Gradients}
\usage{
calc_loss_and_grads(model, layer_contributions)
}
\arguments{
\item{layer_contributions}{}
}
\value{

}
\description{
Calculate Loss & Gradients
}
