% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreamify.R
\name{load_imagenet_model}
\alias{load_imagenet_model}
\title{Load Pre-trained Imagenet Model}
\usage{
load_imagenet_model(...)
}
\arguments{
\item{...}{Other parameters to `keras::application_inception_v3`}
}
\value{
An Inception V3 model, with weights pre-trained on ImageNet.
}
\description{
Load Pre-trained Imagenet Model
}
