% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{pad_and_resize_img}
\alias{pad_and_resize_img}
\title{Pad & Resize an Image}
\usage{
pad_and_resize_img(image_path, size_x = 300, size_y = 300, new_path = NULL)
}
\arguments{
\item{image_path}{hi}

\item{size_x}{bye}

\item{size_y}{:)}

\item{new_path}{die}
}
\value{

}
\description{
Function to pad and resize an image to 300x300, this function currently assumes
that the image is longer than it wide and pads the x axes so that it is as long as wide. Subsequently,
the image is resized.
}
